move things around, var decay config, official ensemble
